fix minor things for strict compiler settings (#331)
In the hunt for some ugly non-reproducibility bug, I used strict
compiler settings with `add_compile_options(-Wall -Wextra -Werror
-Wuninitialized)`

This PR cleans unused variables and incorrect types that don't have any
effect but prevent the usage of strict compiler settings.

+++ include/AdePT/core/AdePTTransport.h
@@ -83,7 +83,7 @@ private:
   int fMaxBatch{0};                                    ///< Max batch size for allocating GPU memory
   int fNumVolumes{0};                                  ///< Total number of active logical volumes
   int fNumSensitive{0};                                ///< Total number of sensitive volumes
-  int fBufferThreshold{20};                            ///< Buffer threshold for flushing AdePT transport buffer
+  size_t fBufferThreshold{20};                         ///< Buffer threshold for flushing AdePT transport buffer
   int fDebugLevel{1};                                  ///< Debug level
   int fCUDAStackLimit{0};                              ///< CUDA device stack limit
   GPUstate *fGPUstate{nullptr};                        ///< CUDA state placeholder
